backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
alembic
python-jose
passlib[bcrypt]==1.7.4
pydantic
pydantic-settings
email-validator
bcrypt<4
